Feat(Signing UX): unify transaction details
