--- apps/web/src/components/transactions/TxDetails/Summary/styles.module.css
@@ -1,4 +1,0 @@
-.buttonExpand {
-  margin-top: 8px;
-  padding: 0;
-}
