+++ apps/web/src/components/transactions/TxDetails/index.tsx
@@ -123,7 +123,7 @@ const TxDetailsBlock = ({ txSummary, txDetails }: TxDetailsProps): ReactElement 
         <div className={css.txSummary}>
           {isUntrusted && !isPending && <UnsignedWarning />}
 
-          <Summary txDetails={txDetails} txData={txDetails.txData} txInfo={txDetails.txInfo} isTxDetailsPreview />
+          <Summary txDetails={txDetails} txData={txDetails.txData} txInfo={txDetails.txInfo} />
         </div>
 
         {(isMultiSendTxInfo(txDetails.txInfo) || isOrderTxInfo(txDetails.txInfo)) && (
