+++ apps/web/src/components/tx-flow/flows/AddOwner/index.tsx
@@ -42,7 +42,7 @@ const FlowInner = ({ defaultValues }: { defaultValues: AddOwnerFlowProps }) => {
       },
       {
         txLayoutProps: { title: 'Confirm transaction details', fixedNonce: true },
-        content: <ConfirmTxDetails key={2} onSubmit={() => {}} showMethodCall />,
+        content: <ConfirmTxDetails key={2} onSubmit={() => {}} />,
       },
     ],
     [nextStep, data],
