+++ apps/web/src/components/tx-flow/flows/ChangeThreshold/index.tsx
@@ -39,7 +39,7 @@ const ChangeThresholdFlow = () => {
       },
       {
         txLayoutProps: { title: 'Confirm transaction details', fixedNonce: true },
-        content: <ConfirmTxDetails key={2} onSubmit={() => {}} showMethodCall />,
+        content: <ConfirmTxDetails key={2} onSubmit={() => {}} />,
       },
     ],
     [nextStep, data],
