+++ apps/web/src/features/spaces/components/Dashboard/styles.module.css
