+++ apps/web/src/features/spaces/components/SpaceBreadcrumbs/styles.module.css
@@ -1,5 +1,5 @@
 .spaceName {
-    display: flex;
-    gap: 4px;
-    align-items: center;
+  display: flex;
+  gap: 4px;
+  align-items: center;
 }
